+++ src/Bexio/Resource/Note.php
@@ -4,67 +4,52 @@ namespace Bexio\Resource;
 
 use Bexio\Bexio;
 
-/**
- * Class Note
- */
 class Note extends Bexio
 {
     /**
      * Gets all Notes
-     *
-     * @return array
      */
-    public function getNotes(array $params = [])
+    public function getNotes(array $params = []): mixed
     {
         return $this->client->get('note', $params);
     }
 
     /**
      * Search for notes
-     *
-     * @return mixed
      */
-    public function searchNotes(array $params = [], array $queryParams = [])
+    public function searchNotes(array $params = [], array $queryParams = []): mixed
     {
         return $this->client->post('note/search', $params, $queryParams);
     }
 
     /**
      * Get specific note
-     *
-     * @return mixed
      */
-    public function getNote(int $id)
+    public function getNote(int $id): mixed
     {
         return $this->client->get("note/$id");
     }
 
     /**
      * Add new note
-     *
-     * @return mixed
      */
-    public function createNote(array $params = [])
+    public function createNote(array $params = []): mixed
     {
         return $this->client->post('note', $params);
     }
 
     /**
      * Edit note
-     *
-     * @return mixed
      */
-    public function editNote(int $id, array $params = [])
+    public function editNote(int $id, array $params = []): mixed
     {
         return $this->client->post("note/$id", $params);
     }
 
     /**
      * Delete specific note
-     *
-     * @return mixed
      */
-    public function deleteNote(int $id)
+    public function deleteNote(int $id): mixed
     {
         return $this->client->delete("note/$id");
     }