+++ src/Bexio/Resource/Other.php
@@ -4,67 +4,52 @@ namespace Bexio\Resource;
 
 use Bexio\Bexio;
 
-/**
- * Class Other
- */
 class Other extends Bexio
 {
     /**
      * Fetch a list of company profiles
-     *
-     * @return array
      */
-    public function getCompanyProfiles()
+    public function getCompanyProfiles(): mixed
     {
         return $this->client->get('company_profile');
     }
 
     /**
      * Get company profile
-     *
-     * @return mixed
      */
-    public function getCompanyProfile(int $id)
+    public function getCompanyProfile(int $id): mixed
     {
         return $this->client->get("company_profile/$id");
     }
 
     /**
      * Get available countries
-     *
-     * @return mixed
      */
-    public function getCountries(array $params = [])
+    public function getCountries(array $params = []): mixed
     {
         return $this->client->get('country', $params);
     }
 
     /**
      * Get available languages
-     *
-     * @return mixed
      */
-    public function getLanguages(array $params = [])
+    public function getLanguages(array $params = []): mixed
     {
         return $this->client->get('language', $params);
     }
 
     /**
      * Get available payment types
-     *
-     * @return mixed
      */
-    public function getPaymentTypes(array $params = [])
+    public function getPaymentTypes(array $params = []): mixed
     {
         return $this->client->get('payment_type', $params);
     }
 
     /**
      * Get available units
-     *
-     * @return mixed
      */
-    public function getUnits(array $params = [])
+    public function getUnits(array $params = []): mixed
     {
         return $this->client->get('unit', $params);
     }